Remove comments around CFBundleShortVersionString
CFBundleShortVersionString was commented out. This causes a problem
when tools like [munki](https://github.com/munki/munki) are used to
deploy the software on multiple machines. Munki uses the information
found in CFBundleShortVersionString to decide whether or not a
computer is running the latest version of a given software. A valid
entry is therefore required.